harden wire-boundary reads and align affinity docs
RingBufferReceiver.Read and the stream chunk path both trusted
peer-provided sizes: a corrupt WriteOffset (available > capacity) or a
hostile PayloadSize >= 2^31 (negative after signed int conversion on
32-bit) could drive a copy/slice past the backing array. Clamp both with
unsigned comparisons and pin the behaviour with wire_guard_test.go.

Also refresh the AffinityMode doc comments in client.go/direct.go to
describe the AffinityAuto zero-value default (chipset-aware since v0.8.2)
instead of the stale AffinityNone-default wording.

+++ go/stream.go
@@ -169,7 +169,10 @@ func NewStreamReassembler(onStream StreamHandler, fallback func(req []byte, resp
 			payloadSize := binary.LittleEndian.Uint32(req[12:16])
 
 			headerLen := int(unsafe.Sizeof(ChunkHeader{}))
-			if len(req) < headerLen+int(payloadSize) {
+			// Unsigned arithmetic: on 32-bit builds int(payloadSize) of a
+			// hostile value >= 2^31 would go negative and slip past a signed
+			// comparison, panicking at the slice expression below.
+			if uint64(len(req)) < uint64(headerLen)+uint64(payloadSize) {
 				return 0, MsgTypeSystemError
 			}
